requirements.txt
huggingface_hub==0.25.2
omegaconf 
pytorch-lightning 
torch 
torchvision 
pydantic
ftfy 
transformers 
diffusers 
fire 
kornia 
accelerate
lovely-numpy 
lovely-tensors
einops 
clip
carvekit 
taming-transformers-rom1504